authentication/views.py
from rest_framework.generics import CreateAPIView
from django.contrib.auth.models import User
from .serializers import RegisterSerializer
from rest_framework.authtoken.models import Token
from rest_framework.response import Response
from rest_framework.authtoken.views import ObtainAuthToken
from users.models import Profile


class UserRegisterView(CreateAPIView):
    serializer_class = RegisterSerializer
    queryset = User.objects.all()


class UserLoginView(ObtainAuthToken):
    def post(self, request, *args, **kwargs):
        serializer = self.serializer_class(data=request.data,
                                           context={'request': request})
        serializer.is_valid(raise_exception=True)
        user = serializer.validated_data['user']
        token, created = Token.objects.get_or_create(user=user)
        return Response({
            'token': token.key,
            'user_id': user.pk,
            'profile_id': self.get_profile_id(user)
        })

    def get_profile_id(self, user):
        return Profile.objects.get(user=user).id
