requirements.txt
asgiref==3.7.2
Django==5.0.2
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
phonenumbers==8.13.29
progressbar2==4.3.2
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
Unidecode==1.3.8
